feat(widget-chat): store conversation
