+++ packages/widget-chat/src/store.js
@@ -2,7 +2,7 @@ import {applyMiddleware, combineReducers, compose, createStore} from 'redux';
 
 import thunk from 'redux-thunk';
 import createLogger from 'redux-logger';
-import user from './reducers';
+import {user, conversation} from './reducers';
 import sparkReducer from './modules/redux-spark/reducers';
 
 const devtools = window.devToolsExtension || (() => (noop) => noop);
@@ -23,6 +23,7 @@ const enhancers = [
 export default createStore(
   combineReducers({
     user,
+    conversation,
     spark: sparkReducer
   }),
   compose(...enhancers)
